+++ lib/src/ui/components/boxes/questionnaire_choice_item.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 
 import '../../layout/inherited_questionnaire_renderer.dart';
-import '../../../core/utils/fhir_renderer_questionnaire_response_utils.dart';
 import 'base_decorator.dart';
 import '../questionnaire_base_item.dart';
-import '../../../core/mixins/choice_value_mixin.dart';
+import '../../../core/mixins/choice_base_mixin.dart';
+import '../../../core/mixins/choice_widget_builder_mixin.dart';
 
 class QuestionnaireChoiceItem extends QuestionnaireBaseItem
-    with ChoiceValueMixin {
+    with ChoiceBaseMixin, ChoiceWidgetBuilderMixin {
   const QuestionnaireChoiceItem({
     required super.questionnaireItem,
     required super.index,
@@ -17,42 +17,22 @@ class QuestionnaireChoiceItem extends QuestionnaireBaseItem
 
   @override
   Widget buildQuestionnaireItem(BuildContext context) {
+    final repeats = questionnaireItem.repeats?.valueBoolean ?? false;
+
+    final selectedResponseItem = findQuestionnaireResponseItem(
+      InheritedQuestionnaireRenderer.of(context).questionnaireResponse,
+      itemLinkId,
+    );
+
     return BaseDecorator(
       roundBottomBorder: isLastItem,
       title: itemTextTitle,
-      children: questionnaireItem.answerOption?.map((answerOption) {
-            String displayValue =
-                answerOption.valueCoding?.display?.valueString ??
-                    answerOption.valueCoding?.code?.valueString ??
-                    "--";
-
-            final selectedResponseItem = findQuestionnaireResponseItem(
-              InheritedQuestionnaireRenderer.of(context).questionnaireResponse,
-              itemLinkId,
-            );
-
-            String? selectedValue = getInitialOrSelectedValue(
-                selectedResponseItem, questionnaireItem);
-
-            return RadioListTile(
-              value: selectedValue,
-              groupValue: answerOption.valueCoding?.code?.valueString,
-              onChanged: (v) => InheritedQuestionnaireRenderer.of(
-                context,
-              ).onResponseChanged(
-                FhirRendererQuestionnaireResponseUtils
-                    .setAnswerOptionInQuestionnaireResponse(
-                  InheritedQuestionnaireRenderer.of(
-                    context,
-                  ).questionnaireResponse,
-                  questionnaireItem,
-                  answerOption,
-                ),
-              ),
-              title: Text(displayValue),
-            );
-          }).toList() ??
-          [],
+      children: buildOptionWidgets(
+        context: context,
+        questionnaireItem: questionnaireItem,
+        selectedResponseItem: selectedResponseItem,
+        repeats: repeats,
+      ),
     );
   }
 }